Move and document Resource constructors
